+++ app/src/main/java/com/bobble/bobblesampleapp/database/repository/StickersRepository.java
@@ -25,7 +25,7 @@ public class StickersRepository {
         return (getstickerDao(context).count() == 0);
     }
 
-    public static void deleteStikerWithId(Context context, int id) {
+    public static void deleteStikerWithId(Context context, long id) {
         getstickerDao(context).delete(getstickerDaoForId(context, id));
     }
 
@@ -33,11 +33,10 @@ public class StickersRepository {
         return getstickerDao(context).loadAll();
     }
 
-    public static Sticker getstickerDaoForId(Context context, int id) {
+    public static Sticker getstickerDaoForId(Context context, long id) {
         return getstickerDao(context).load(id);
     }
     public static StickerDao getstickerDao(Context c) {
         return ((BobbleSampleApp) c.getApplicationContext()).getDaoSession().getStickerDao();
     }
-
 }
